start to divide the parsing logic from the generation logic
